pyproject.toml
[project]
name = "sceodesic"
version = "0.0.1"
dependencies = [
    "setuptools",
    "numpy", #< 1.24",
    "pandas", #< 1.4", 
    "scipy", #< 2.0", 
    "scanpy", #< 2.0", 
    "anndata", #< 0.9.0",
    "fbpca",
    "scikit-learn", #< 1.2",
    "PyYaml" #< 6.1"
]
authors = [
  { name="Sinan Ozbay", email="sinan.ozbay@duke.edu" },
  { name="Aditya Parekh", email="aditya.parekh@duke.edu" }
]
description = "Generate sceodesic embeddings from an input scRNA-seq dataset."
readme = "README.md"
#requires-python = ">=3.9,<3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/rohitsinghlab/sceodesic"
"Bug Tracker" = "https://github.com/rohitsinghlab/sceodesic/issues"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["*test*"]
